Workaround for premature using of unloaded obj
An IntangibleObject that hadn't been initialized yet was having its
baselines sent out, causing an objectMutex null error.

No idea why but from the prints it was clear that no IntangibleObject
had yet been initialized when the server attempted to send Intangible
baselines.

It's really a hacky patchfix so can't be certain the issue won't
manifest itself in other ways.