Fixed ConcurrentModification with observers
A concurrent modification exception was occuring with the observer list (very rarely to the extent I only saw it once but still shouldn't happen!).

I suspect this was occuring because, although notifyObservers is synchronized (with objectMutex and not the list itself as java docs says to), people who use getObservers() will still be iterating through the observers at the same time as notifyObservers is.  The easiest solution is probably to just have getObservers() return a clone.

I'm not sure if it's safe to replace the "objectMutex" with "observers" but can't currently see a reason why it would be a problem for people to modify other object variables while an already-produced byte array is being sent.